pyproject.toml
[project]
name = "gdg-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.25.1",
    "litellm>=1.81.13",
]
